requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.0
behave==1.2.6
bidict==0.22.0
cfgv==3.3.1
click==8.1.3
coverage==6.4.4
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
flask-restx==1.0.3
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
identify==2.5.5
importlib-metadata==4.12.0
importlib-resources==5.10.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
nodeenv==1.7.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.1.3
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.41
toml==0.10.2
tomli==2.0.1
virtualenv==20.16.5
Werkzeug==2.2.2
zipp==3.8.1
